Rework strata UI: 3x2 grid, continuous rate slider, per-stratum VU meters
Reorganize strata from 6 stacked rows into a 3x2 grid layout. Replace the
discrete F/M/S segmented control with a continuous rate slider using exponential
duration interpolation. Add per-stratum level metering with vertical VU bars
that animate during playback. Move toggle switch next to range label and remove
the NoteIndicator. Add run.ps1 for building without Maven.

+++ src/main/java/org/delightofcomposition/markov/Cue.java
@@ -10,13 +10,13 @@ public class Cue {
     private final Scale scale;
     private final boolean[] active;     // 6 strata
     private final double[] amplitudes;  // 6 strata
-    private final int[] rateLevels;     // 6 strata
+    private final double[] rateLevels;  // 6 strata (continuous 0.0-2.0)
     private final double reverbRoom;
     private final double reverbMix;
     private final double reverbDamp;
 
     public Cue(String name, double[][] matrix, Scale scale,
-               boolean[] active, double[] amplitudes, int[] rateLevels,
+               boolean[] active, double[] amplitudes, double[] rateLevels,
                double reverbRoom, double reverbMix, double reverbDamp) {
         this.name = name;
         this.matrix = deepCopy(matrix);
@@ -34,7 +34,7 @@ public class Cue {
     public Scale getScale() { return scale; }
     public boolean[] getActive() { return active.clone(); }
     public double[] getAmplitudes() { return amplitudes.clone(); }
-    public int[] getRateLevels() { return rateLevels.clone(); }
+    public double[] getRateLevels() { return rateLevels.clone(); }
     public double getReverbRoom() { return reverbRoom; }
     public double getReverbMix() { return reverbMix; }
     public double getReverbDamp() { return reverbDamp; }
